requirements.txt
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
requests==2.32.3
soupsieve==2.6
urllib3==2.3.0
pysimplegui==5.0.8
cohere==5.13.12
flask==3.1.0
beautifulsoup4
certifi
charset-normalizer
cohere
flask
idna
pysimplegui
requests
soupsieve
urllib3
python-dotenv
